interface, class モジュール化
User,RoomHash -> interface
Room,RoomFactory -> model
