+++ src/interface/roomHash.ts
@@ -1,0 +1,12 @@
+import { User } from '../interface/user'
+
+export interface RoomHash {
+  name: string
+  password: string
+  entryNum: number
+  mode: string
+  level: string
+  round: number
+  participants: User[]
+  link: string
+}
