+++ src/interface/user.ts
@@ -1,0 +1,9 @@
+export interface User {
+  name: string
+  mail: string
+  password: string
+  img: string
+  profile: string
+  twitterAccount: string
+  login: boolean
+}
